requirements.txt
python-telegram-bot==20.7
yfinance==0.2.28
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
ta>=0.10.0
python-dotenv>=1.0.0
schedule>=1.2.0
aiohttp>=3.9.0
sqlite3
datetime
json
logging 